make OfficeDevTreeViewManager
