+++ packages/vscode-extension/src/treeview/treeViewManager.ts
@@ -47,33 +47,6 @@ class TreeViewManager {
     context.subscriptions.push(...disposables);
   }
 
-  public registerOfficeDevTreeViews(context: vscode.ExtensionContext): void {
-    const disposables: vscode.Disposable[] = [];
-
-    this.registerOfficeDevTreeViewItems(
-      disposables,
-      "teamsfx-officedev-development",
-      this.getOfficeDevelopmentCommands
-    );
-    this.registerOfficeDevTreeViewItems(
-      disposables,
-      "teamsfx-officedev-lifecycle",
-      this.getOfficeLifecycleCommands
-    );
-    this.registerOfficeDevTreeViewItems(
-      disposables,
-      "teamsfx-officedev-utility",
-      this.getOfficeUtilityCommands
-    );
-    this.registerOfficeDevTreeViewItems(
-      disposables,
-      "teamsfx-officedev-help-and-feedback",
-      this.getOfficeHelperCommands
-    );
-
-    context.subscriptions.push(...disposables);
-  }
-
   public async updateTreeViewsByContent(removeProjectRelatedCommands = false): Promise<void> {
     let isTeamsApp = false;
     const manifestRes = await manifestUtils.readAppManifest(workspaceUri?.fsPath || "");
@@ -365,150 +338,6 @@ class TreeViewManager {
     );
     this.storeCommandsIntoMap(helpCommand);
     this.treeviewMap.set("teamsfx-help-and-feedback", helpProvider);
-  }
-
-  private registerOfficeDevTreeViewItems(
-    disposables: vscode.Disposable[],
-    treeViewId: string,
-    getCommandsFunc: () => TreeViewCommand[]
-  ) {
-    const commands = getCommandsFunc();
-    const provider = new CommandsTreeViewProvider(commands);
-    disposables.push(vscode.window.registerTreeDataProvider(treeViewId, provider));
-    this.storeCommandsIntoMap(commands);
-    this.treeviewMap.set(treeViewId, provider);
-  }
-
-  private getOfficeDevelopmentCommands(): TreeViewCommand[] {
-    return [
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.createOfficeAddInTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.createOfficeAddInDescription"),
-        "fx-extension.create",
-        undefined,
-        { name: "new-folder", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.samplesTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.samplesDescription"),
-        "fx-extension.openSamples",
-        undefined,
-        { name: "library", custom: false },
-        TreeCategory.GettingStarted
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.checkAndInstallDependenciesTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.checkAndInstallDependenciesDescription"),
-        "fx-extension.installDependency",
-        undefined,
-        { name: "check-all", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeDevLocalDebugTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.officeDevLocalDebugDescription"),
-        "fx-extension.localdebug",
-        undefined,
-        { name: "debug-alt", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.stopDebugTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.stopDebugDescription"),
-        "fx-extension.stopDebugging",
-        undefined,
-        { name: "debug-stop", custom: false }
-      ),
-    ];
-  }
-
-  private getOfficeLifecycleCommands(): TreeViewCommand[] {
-    return [
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeDevDeployTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.officeDevDeployDescription"),
-        "fx-extension.officeDevDeploy",
-        undefined,
-        { name: "cloud-upload", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.publishAppSourceTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.publishAppSourceDescription"),
-        "fx-extension.publishToAppSource",
-        undefined,
-        { name: "teamsfx-developer-portal", custom: false }
-      ),
-    ];
-  }
-
-  private getOfficeUtilityCommands(): TreeViewCommand[] {
-    const officeUtilityCommands = [
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.validateManifestTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.validateManifestDescription"),
-        "fx-extension.validateApplication",
-        undefined,
-        {
-          name: "beaker",
-          custom: false,
-        }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.editManifestTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.editManifestDescription"),
-        "fx-extension.editManifest",
-        undefined,
-        { name: "note", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.generateManifestGUIDTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.generateManifestGUIDDescription"),
-        "fx-extension.generateManifestGUID",
-        undefined,
-        { name: "gripper", custom: false }
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.officePartnerCenterTitle"),
-        localize(
-          "teamstoolkit.commandsTreeViewProvider.officeAddIn.officePartnerCenterDescription"
-        ),
-        "fx-extension.officePartnerCenter",
-        undefined,
-        { name: "unfold", custom: false },
-        TreeCategory.GettingStarted
-      ),
-    ];
-
-    return officeUtilityCommands;
-  }
-
-  private getOfficeHelperCommands(): TreeViewCommand[] {
-    const helpCommand = [
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.documentationTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.documentationDescription"),
-        "fx-extension.openOfficeDevDocument",
-        undefined,
-        { name: "book", custom: false },
-        TreeCategory.GettingStarted
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.getStartedTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.officeAddIn.getStartedDescription"),
-        "fx-extension.openGetStarted",
-        undefined,
-        { name: "symbol-event", custom: false },
-        TreeCategory.GettingStarted
-      ),
-      new TreeViewCommand(
-        localize("teamstoolkit.commandsTreeViewProvider.reportIssuesTitle"),
-        localize("teamstoolkit.commandsTreeViewProvider.reportIssuesDescription"),
-        "fx-extension.openOfficeDevReportIssues",
-        undefined,
-        { name: "github", custom: false },
-        TreeCategory.Feedback
-      ),
-    ];
-
-    return helpCommand;
   }
 
   private storeCommandsIntoMap(commands: TreeViewCommand[]) {
